added local state to store entry id and file offset position

+++ ILogger.hpp
@@ -8,9 +8,9 @@ namespace Logger
     {
     public:
         virtual ~ILogger() {}
-        virtual uint64_t Append(std::vector<uint8_t> data) = 0;
-        virtual uint64_t GetPosition() const = 0;
-        virtual void Truncate(uint64_t position) = 0;
-        virtual void Replay(uint64_t position, std::function<void(std::vector<uint8_t>)> callback) const = 0;
+        virtual uint32_t Append(std::vector<uint8_t> data) = 0;
+        virtual uint32_t GetPosition() const = 0;
+        virtual void Truncate(uint32_t position) = 0;
+        virtual void Replay(uint32_t position, std::function<void(std::vector<uint8_t>)> callback) const = 0;
     };
 }
